Added some style to the menu profile, refactored and link to the user profile
