+++ src/pwa/service-worker.js
@@ -4,7 +4,7 @@ import { StaleWhileRevalidate } from "workbox-strategies"
 import { setCacheNameDetails } from 'workbox-core'
 
 const CACHE_NAME = 'crawlearcom'
-const CACHE_VERSION = '1.0.37'
+const CACHE_VERSION = '1.0.38'
 //const PREVIOUS_CACHE_VERSION = '1.0.34'
 //crawlearcom-crawlearcom-precache-1.0.21
